fixed profiling when set to false (default); renamed decorator and file name for contexts to be cleaner
