+++ src/domain/challenge/challenge.resolver.fields.ts
@@ -12,7 +12,7 @@ import { User } from '../user/user.entity';
 import { UserGroupService } from '../user-group/user-group.service';
 import { ChallengeService } from './challenge.service';
 import { Opportunity } from '../opportunity/opportunity.entity';
-import { Measure } from '../../utils/logging/logging.profile.decorator';
+import { Profiling } from '../../utils/logging/logging.profiling.decorator';
 
 @Resolver(() => Challenge)
 export class ChallengeResolverFields {
@@ -30,7 +30,7 @@ export class ChallengeResolverFields {
     nullable: true,
     description: 'Groups of users related to a challenge.',
   })
-  @Measure.api
+  @Profiling.api
   async groups(@Parent() challenge: Challenge) {
     const groups = await this.challengeService.loadGroups(challenge);
     return groups;
@@ -45,7 +45,7 @@ export class ChallengeResolverFields {
     nullable: true,
     description: 'The set of opportunities within this challenge.',
   })
-  @Measure.api
+  @Profiling.api
   async opportunities(@Parent() challenge: Challenge) {
     const opportunities = await this.challengeService.loadOpportunities(
       challenge
@@ -62,7 +62,7 @@ export class ChallengeResolverFields {
     nullable: true,
     description: 'All users that are contributing to this challenge.',
   })
-  @Measure.api
+  @Profiling.api
   async contributors(@Parent() challenge: Challenge) {
     const group = await this.userGroupService.getGroupByName(
       challenge,
