+++ src/domain/challenge/challenge.resolver.mutations.ts
@@ -14,7 +14,7 @@ import { IChallenge } from './challenge.interface';
 import { ChallengeService } from './challenge.service';
 import { OpportunityInput } from '../opportunity/opportunity.dto';
 import { Opportunity } from '../opportunity/opportunity.entity';
-import { Measure } from '../../utils/logging/logging.profile.decorator';
+import { Profiling } from '../../utils/logging/logging.profiling.decorator';
 
 @Resolver()
 export class ChallengeResolverMutations {
@@ -30,7 +30,7 @@ export class ChallengeResolverMutations {
   @Mutation(() => UserGroup, {
     description: 'Creates a new user group for the challenge with the given id',
   })
-  @Measure.api
+  @Profiling.api
   async createGroupOnChallenge(
     @Args({ name: 'challengeID', type: () => Float }) challengeID: number,
     @Args({ name: 'groupName', type: () => String }) groupName: string
@@ -51,7 +51,7 @@ export class ChallengeResolverMutations {
     description:
       'Creates a new Opportunity for the challenge with the given id',
   })
-  @Measure.api
+  @Profiling.api
   async createOpportunityOnChallenge(
     @Args({ name: 'challengeID', type: () => Float }) challengeID: number,
     @Args({ name: 'opportunityData', type: () => OpportunityInput })
@@ -70,7 +70,7 @@ export class ChallengeResolverMutations {
     description:
       'Updates the specified Challenge with the provided data (merge)',
   })
-  @Measure.api
+  @Profiling.api
   async updateChallenge(
     @Args({ name: 'challengeID', type: () => Float }) challengeID: number,
     @Args('challengeData') challengeData: ChallengeInput
@@ -91,7 +91,7 @@ export class ChallengeResolverMutations {
     description:
       'Adds the user with the given identifier as a member of the specified challenge',
   })
-  @Measure.api
+  @Profiling.api
   async addUserToChallenge(
     @Args('userID') userID: number,
     @Args('challengeID') challengeID: number
