+++ src/domain/ecoverse/ecoverse.resolver.mutations.ts
@@ -3,7 +3,7 @@ import { Resolver } from '@nestjs/graphql';
 import { Args, Mutation } from '@nestjs/graphql/dist/decorators';
 import { GqlAuthGuard } from '../../utils/authentication/graphql.guard';
 import { Roles } from '../../utils/decorators/roles.decorator';
-import { Measure } from '../../utils/logging/logging.profile.decorator';
+import { Profiling } from '../../utils/logging/logging.profiling.decorator';
 import { ChallengeInput } from '../challenge/challenge.dto';
 import { Challenge } from '../challenge/challenge.entity';
 import { IChallenge } from '../challenge/challenge.interface';
@@ -37,7 +37,7 @@ export class EcoverseResolverMutations {
   @Mutation(() => UserGroup, {
     description: 'Creates a new user group at the ecoverse level',
   })
-  @Measure.api
+  @Profiling.api
   async createGroupOnEcoverse(
     @Args({ name: 'groupName', type: () => String }) groupName: string
   ): Promise<IUserGroup> {
@@ -50,7 +50,7 @@ export class EcoverseResolverMutations {
   @Mutation(() => Ecoverse, {
     description: 'Updates the Ecoverse with the provided data',
   })
-  @Measure.api
+  @Profiling.api
   async updateEcoverse(
     @Args('ecoverseData') ecoverseData: EcoverseInput
   ): Promise<IEcoverse> {
@@ -67,7 +67,7 @@ export class EcoverseResolverMutations {
     description:
       'Creates a new user as a member of the ecoverse, including an account if enabled',
   })
-  @Measure.api
+  @Profiling.api
   async createUser(@Args('userData') userData: UserInput): Promise<IUser> {
     const user = await this.ecoverseService.createUser(userData);
     return user;
@@ -79,7 +79,7 @@ export class EcoverseResolverMutations {
     description:
       'Creates a new template for the population of entities within tis ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async createTemplate(
     @Args('templateData') templateData: TemplateInput
   ): Promise<ITemplate> {
@@ -96,7 +96,7 @@ export class EcoverseResolverMutations {
     description:
       'Creates a new user as a member of the ecoverse, without an account',
   })
-  @Measure.api
+  @Profiling.api
   async createUserProfile(
     @Args('userData') userData: UserInput
   ): Promise<IUser> {
@@ -112,7 +112,7 @@ export class EcoverseResolverMutations {
   @Mutation(() => Boolean, {
     description: 'Removes the specified user from the ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async removeUser(@Args('userID') userID: number): Promise<boolean> {
     const success = await this.ecoverseService.removeUser(userID);
     return success;
@@ -123,7 +123,7 @@ export class EcoverseResolverMutations {
   @Mutation(() => Challenge, {
     description: 'Creates a new challenge and registers it with the ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async createChallenge(
     @Args('challengeData') challengeData: ChallengeInput
   ): Promise<IChallenge> {
@@ -138,7 +138,7 @@ export class EcoverseResolverMutations {
     description:
       'Creates a new organisation and registers it with the ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async createOrganisation(
     @Args('organisationData') organisationData: OrganisationInput
   ): Promise<IOrganisation> {
