+++ src/domain/ecoverse/ecoverse.resolver.queries.ts
@@ -2,7 +2,7 @@ import { Inject, UseGuards } from '@nestjs/common';
 import { Args, Query, Resolver } from '@nestjs/graphql';
 import { GqlAuthGuard } from '../../utils/authentication/graphql.guard';
 import { Roles } from '../../utils/decorators/roles.decorator';
-import { Measure } from '../../utils/logging/logging.profile.decorator';
+import { Profiling } from '../../utils/logging/logging.profiling.decorator';
 import { Challenge } from '../challenge/challenge.entity';
 import { IChallenge } from '../challenge/challenge.interface';
 import { ChallengeService } from '../challenge/challenge.service';
@@ -40,7 +40,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'The name for this ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async name(): Promise<string> {
     return this.ecoverseService.getName();
   }
@@ -49,7 +49,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'The host organisation for the ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async host(): Promise<IOrganisation> {
     return this.ecoverseService.getHost();
   }
@@ -58,7 +58,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'The shared understanding for this ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async context(): Promise<IContext> {
     return this.ecoverseService.getContext();
   }
@@ -72,7 +72,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'The members of this this ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async users(): Promise<IUser[]> {
     return this.ecoverseService.getUsers();
   }
@@ -87,7 +87,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'A particular user, identified by the ID or by email',
   })
-  @Measure.api
+  @Profiling.api
   async user(@Args('ID') id: string): Promise<IUser> {
     const user = await this.userService.getUser(id);
     if (user) return user;
@@ -105,7 +105,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'The members of this this ecoverse filtered by list of IDs.',
   })
-  @Measure.api
+  @Profiling.api
   async usersById(
     @Args({ name: 'IDs', type: () => [String] }) ids: string[]
   ): Promise<IUser[]> {
@@ -124,7 +124,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'All groups at the ecoverse level',
   })
-  @Measure.api
+  @Profiling.api
   async groups(): Promise<IUserGroup[]> {
     const groups = await this.ecoverseService.getGroups();
     return groups;
@@ -139,7 +139,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'All groups that have the provided tag',
   })
-  @Measure.api
+  @Profiling.api
   async groupsWithTag(@Args('tag') tag: string): Promise<IUserGroup[]> {
     const groups = await this.ecoverseService.getGroupsWithTag(tag);
     return groups;
@@ -155,7 +155,7 @@ export class EcoverseResolverQueries {
     description:
       'The user group with the specified id anywhere in the ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async group(@Args('ID') id: number): Promise<IUserGroup | undefined> {
     const group = await this.groupService.getGroupByID(id, {
       relations: ['members', 'focalPoint'],
@@ -164,14 +164,14 @@ export class EcoverseResolverQueries {
   }
 
   @Query(() => [Challenge], { nullable: false, description: 'All challenges' })
-  @Measure.api
+  @Profiling.api
   async challenges(): Promise<IChallenge[]> {
     const challenges = await this.ecoverseService.getChallenges();
     return challenges;
   }
 
   @Query(() => [Template], { nullable: false, description: 'All templates' })
-  @Measure.api
+  @Profiling.api
   async templates(): Promise<ITemplate[]> {
     const templates = await this.ecoverseService.getTemplates();
     return templates;
@@ -181,7 +181,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'A particular challenge',
   })
-  @Measure.api
+  @Profiling.api
   async challenge(@Args('ID') id: number): Promise<IChallenge | undefined> {
     return await this.challengeService.getChallengeByID(id);
   }
@@ -190,7 +190,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'All organisations',
   })
-  @Measure.api
+  @Profiling.api
   async organisations(): Promise<IOrganisation[]> {
     const organisations = await this.ecoverseService.getOrganisations();
     return organisations;
@@ -200,7 +200,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'A particular organisation',
   })
-  @Measure.api
+  @Profiling.api
   async organisation(
     @Args('ID') id: number
   ): Promise<IOrganisation | undefined> {
@@ -211,7 +211,7 @@ export class EcoverseResolverQueries {
     nullable: false,
     description: 'The tagset associated with this Ecoverse',
   })
-  @Measure.api
+  @Profiling.api
   async tagset(): Promise<ITagset> {
     return await this.ecoverseService.getTagset();
   }
