+++ src/domain/organisation/organisation.resolver.fields.ts
@@ -10,7 +10,7 @@ import {
 import { Organisation } from './organisation.entity';
 import { User } from '../user/user.entity';
 import { UserGroupService } from '../user-group/user-group.service';
-import { Measure } from '../../utils/logging/logging.profile.decorator';
+import { Profiling } from '../../utils/logging/logging.profiling.decorator';
 
 @Resolver(() => Organisation)
 export class OrganisationResolverFields {
@@ -25,7 +25,7 @@ export class OrganisationResolverFields {
     nullable: true,
     description: 'Groups defined on this organisation.',
   })
-  @Measure.api
+  @Profiling.api
   async groups(@Parent() organisation: Organisation) {
     const groups = await organisation.groups;
     if (!groups)
@@ -42,7 +42,7 @@ export class OrganisationResolverFields {
     nullable: true,
     description: 'Users that are contributing to this organisation.',
   })
-  @Measure.api
+  @Profiling.api
   async contributors(@Parent() organisation: Organisation) {
     const group = await this.userGroupService.getGroupByName(
       organisation,
