+++ src/domain/user-group/user-group.resolver.ts
@@ -12,7 +12,7 @@ import { RestrictedGroupNames, UserGroup } from './user-group.entity';
 import { IUserGroup } from './user-group.interface';
 import { UserGroupService } from './user-group.service';
 import { User } from '../user/user.entity';
-import { Measure } from '../../utils/logging/logging.profile.decorator';
+import { Profiling } from '../../utils/logging/logging.profiling.decorator';
 
 @Resolver(() => UserGroup)
 export class UserGroupResolver {
@@ -27,7 +27,7 @@ export class UserGroupResolver {
     description:
       'Adds the user with the given identifier to the specified user group',
   })
-  @Measure.api
+  @Profiling.api
   async addUserToGroup(
     @Args('userID') userID: number,
     @Args('groupID') groupID: number
@@ -45,7 +45,7 @@ export class UserGroupResolver {
     description:
       'Remove the user with the given identifier to the specified user group',
   })
-  @Measure.api
+  @Profiling.api
   async removeUserFromGroup(
     @Args('userID') userID: number,
     @Args('groupID') groupID: number
@@ -64,7 +64,7 @@ export class UserGroupResolver {
     description:
       'Assign the user with the given ID as focal point for the given group',
   })
-  @Measure.api
+  @Profiling.api
   async assignGroupFocalPoint(
     @Args('userID') userID: number,
     @Args('groupID') groupID: number
@@ -82,7 +82,7 @@ export class UserGroupResolver {
     nullable: true,
     description: 'Remove the focal point for the given group',
   })
-  @Measure.api
+  @Profiling.api
   async removeGroupFocalPoint(
     @Args('groupID') groupID: number
   ): Promise<IUserGroup> {
@@ -91,7 +91,7 @@ export class UserGroupResolver {
   }
 
   @ResolveField('members', () => User)
-  @Measure.api
+  @Profiling.api
   async members(@Parent() group: UserGroup): Promise<User[]> {
     if (!group || !group.membersPopulationEnabled) return [];
